Changed amount fields from decimal to big integer in transactions

+++ database/migrations/2024_03_16_164331_create_accounts_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->timestamps();
 
             $table->string('name')->index();
-            $table->decimal('balance', 13, 4)->default(0.00)->index();
+            $table->bigInteger('balance')->default(0)->index();
             $currencies = array_column(Currency::cases(), 'value');
             $table->enum('currency', $currencies)->default(Currency::USD->name)->index();
             $table->text('description')->nullable();
